Add validation

+++ database/database/src/main/java/team/college/database/service/OrderService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 
 import team.college.database.model.entity.*;
 import team.college.database.model.repository.OrderRepository;
+import team.college.database.model.repository.UserRepository;
 
 @Service
 public class OrderService {
@@ -12,19 +13,44 @@ public class OrderService {
         @Autowired
         private OrderRepository orderRepository;
 
-        public void add(Order order) {
-                orderRepository.save(order);
+        @Autowired
+        private UserRepository userRepository;
+
+        public Boolean add(Order order) {
+                if (order == null) return false;
+                if (order.getTotalCost() == null || order.getTotalCost() <= 0) return false;
+                if (order.getId() != null) return false;
+                if (order.getUser() == null) return false;
+                if (userRepository.findById(order.getUser().getId()).isPresent())
+                {
+                        orderRepository.save(order);
+                        return true;
+                }
+                return false;
         }
 
-        public void update(Order order) {
-                orderRepository.save(order);
+        public Boolean update(Order order) {
+                if (order == null) return false;
+                if (order.getTotalCost() == null || order.getTotalCost() <= 0) return false;
+                if (order.getId() == null) return false;
+                if (orderRepository.findById(order.getId()).isEmpty()) return false;
+                if (order.getUser() == null) return false;
+                if (userRepository.findById(order.getUser().getId()).isPresent())
+                {
+                        orderRepository.save(order);
+                        return true;
+                }
+                return false;
         }
 
-        public void remove(Integer order_id) {
+        public Boolean remove(Integer order_id) {
+                if (orderRepository.findById(order_id).isEmpty()) return false;
                 orderRepository.deleteById(order_id);
+                return true;
         }
 
         public Order get(Integer order_id) {
+                if (orderRepository.findById(order_id).isEmpty()) return null;
                 return orderRepository.findById(order_id).get();
         }
 }
